Split matrix expressions into separate files
