+++ symengine/matrices/identity_matrix.cpp
@@ -1,0 +1,63 @@
+#include <symengine/basic.h>
+#include <symengine/number.h>
+#include <symengine/integer.h>
+#include <symengine/matrices/matrix_expr.h>
+#include <symengine/matrices/identity_matrix.h>
+
+namespace SymEngine
+{
+
+hash_t IdentityMatrix::__hash__() const
+{
+    return n_->__hash__();
+}
+
+bool IdentityMatrix::__eq__(const Basic &o) const
+{
+    return (is_a<IdentityMatrix>(o)
+            && n_->__eq__(*down_cast<const IdentityMatrix &>(o).n_));
+}
+
+int IdentityMatrix::compare(const Basic &o) const
+{
+    SYMENGINE_ASSERT(is_a<IdentityMatrix>(o));
+
+    return n_->compare(*down_cast<const IdentityMatrix &>(o).n_);
+}
+
+vec_basic IdentityMatrix::get_args() const
+{
+    return {n_};
+}
+
+bool IdentityMatrix::is_canonical(const RCP<const Basic> &n) const
+{
+    if (is_a_Number(*n)) {
+        if (is_a<Integer>(*n)) {
+            if (down_cast<const Integer &>(*n).is_negative()) {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+    return true;
+}
+
+RCP<const MatrixExpr> identity_matrix(const RCP<const Basic> &n)
+{
+    if (is_a_Number(*n)) {
+        if (is_a<Integer>(*n)) {
+            if (down_cast<const Integer &>(*n).is_negative()) {
+                throw DomainError(
+                    "Dimension of IdentityMatrix must be nonnegative");
+            }
+        } else {
+            throw DomainError(
+                "Dimension of IdentityMatrix must be a nonnegative integer");
+        }
+    }
+    return make_rcp<const IdentityMatrix>(n);
+}
+
+} // namespace SymEngine
