+++ symengine/matrices/is_real.h
@@ -1,0 +1,15 @@
+#ifndef SYMENGINE_MATRICES_IS_REAL_H
+#define SYMENGINE_MATRICES_IS_REAL_H
+
+#include <symengine/basic.h>
+#include <symengine/assumptions.h>
+#include <symengine/matrices/matrix_expr.h>
+
+namespace SymEngine
+{
+
+tribool is_real(const MatrixExpr &m, const Assumptions *assumptions = nullptr);
+
+}
+
+#endif
