+++ symengine/matrices/matrix_expr.h
@@ -1,0 +1,13 @@
+#ifndef SYMENGINE_MATRICES_MATRIXEXPR_H
+#define SYMENGINE_MATRICES_MATRIXEXPR_H
+
+namespace SymEngine
+{
+
+class MatrixExpr : public Basic
+{
+};
+
+} // namespace SymEngine
+
+#endif
