+++ symengine/matrices/zero_matrix.cpp
@@ -1,0 +1,94 @@
+#include <symengine/number.h>
+#include <symengine/integer.h>
+#include <symengine/matrices/zero_matrix.h>
+
+namespace SymEngine
+{
+
+hash_t ZeroMatrix::__hash__() const
+{
+    hash_t seed = SYMENGINE_ZEROMATRIX;
+    hash_combine<Basic>(seed, *m_);
+    hash_combine<Basic>(seed, *n_);
+    return seed;
+}
+
+bool ZeroMatrix::__eq__(const Basic &o) const
+{
+    return (is_a<ZeroMatrix>(o)
+            && m_->__eq__(*down_cast<const ZeroMatrix &>(o).m_)
+            && n_->__eq__(*down_cast<const ZeroMatrix &>(o).n_));
+}
+
+int ZeroMatrix::compare(const Basic &o) const
+{
+    SYMENGINE_ASSERT(is_a<ZeroMatrix>(o));
+
+    const ZeroMatrix &other = down_cast<const ZeroMatrix &>(o);
+    auto temp = m_->compare(*(other.m_));
+    if (temp != 0) {
+        return temp;
+    } else {
+        return n_->compare(*(other.n_));
+    }
+}
+
+vec_basic ZeroMatrix::get_args() const
+{
+    return {m_, n_};
+}
+
+bool ZeroMatrix::is_canonical(const RCP<const Basic> &m,
+                              const RCP<const Basic> &n) const
+{
+    if (is_a_Number(*m)) {
+        if (is_a<Integer>(*m)) {
+            if (down_cast<const Integer &>(*m).is_negative()) {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+    if (is_a_Number(*n)) {
+        if (is_a<Integer>(*n)) {
+            if (down_cast<const Integer &>(*n).is_negative()) {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+    return true;
+}
+
+RCP<const MatrixExpr> zero_matrix(const RCP<const Basic> &m,
+                                  const RCP<const Basic> &n)
+{
+    if (is_a_Number(*m)) {
+        if (is_a<Integer>(*m)) {
+            if (down_cast<const Integer &>(*m).is_negative()) {
+                throw DomainError(
+                    "Dimension of ZeroMatrix must be nonnegative");
+            }
+        } else {
+            throw DomainError(
+                "Dimension of ZeroMatrix must be a nonnegative integer");
+        }
+    }
+    if (is_a_Number(*n)) {
+        if (is_a<Integer>(*n)) {
+            if (down_cast<const Integer &>(*n).is_negative()) {
+                throw DomainError(
+                    "Dimension of ZeroMatrix must be nonnegative");
+            }
+        } else {
+            throw DomainError(
+                "Dimension of ZeroMatrix must be a nonnegative integer");
+        }
+    }
+
+    return make_rcp<const ZeroMatrix>(m, n);
+}
+
+} // namespace SymEngine
