+++ symengine/tests/matrix/test_matrixexpr.cpp
@@ -19,7 +19,10 @@ using SymEngine::immutable_dense_matrix;
 using SymEngine::ImmutableDenseMatrix;
 using SymEngine::integer;
 using SymEngine::is_a;
+using SymEngine::is_false;
+using SymEngine::is_indeterminate;
 using SymEngine::is_real;
+using SymEngine::is_symmetric;
 using SymEngine::is_true;
 using SymEngine::make_rcp;
 using SymEngine::matrix_add;
